fixed DB n shie

+++ TheMäklers/Data/Models/Housing.cs
@@ -6,14 +6,14 @@
         public string Category { get; set; }
         public string Address { get; set; }
         public string Municipality { get; set; }
-        public decimal InitialPrice { get; set; }
+        public double InitialPrice { get; set; }
         public double LivingArea { get; set; }
         public double AdditionalArea { get; set; }
         public double PlotArea { get; set; }
         public string Description { get; set; }
         public int NumberOfRooms { get; set; }
-        public decimal MonthlyFee { get; set; }
-        public decimal AnnualOperatingCost { get; set; }
+        public double MonthlyFee { get; set; }
+        public double AnnualOperatingCost { get; set; }
         public int YearBuilt { get; set; }
         public List<string> Images { get; set; }
